verlaufsdaten sind jetzt richtig
Former-commit-id: 10f59d8ee629d6e1ca345edaf201365be45dd1a6 [formerly 040a4ff79b4e1c4e3b3cb7d8a48e71fcfbd5402f] [formerly 040a4ff79b4e1c4e3b3cb7d8a48e71fcfbd5402f [formerly 040a4ff79b4e1c4e3b3cb7d8a48e71fcfbd5402f [formerly 040a4ff79b4e1c4e3b3cb7d8a48e71fcfbd5402f [formerly c378d1a42748beacfccb3db01af4240b8e850dc6]]]]
Former-commit-id: 56f85d100174d63ea985fd50685eea916a62c31a
Former-commit-id: 6b748ab7796db6dab8e0faa9e8c8b8c45910521e
Former-commit-id: c54176f3d152243723204174a19a565c887ed71d
Former-commit-id: e43d8d36186085f2c424b1d29ac7e61705ecd881

+++ src/UI.m
@@ -179,7 +179,7 @@ classdef UI < handle
                     if isfield(reader.data, 'sisa')
                         % open a SiSa tab
                         this.modes{1} = InvivoMode(this, double(reader.data.sisa.data),...
-                                                         double(reader.data.sisa.verlauf), reader);
+                                                         reader.data.sisa.verlauf, reader);
                     end
                     if isfield(reader.data, 'fluo')
                         % open a fluorescence tab
